% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\docType{methods}
\name{N}
\alias{N}
\alias{N,PairCoefficients-method}
\title{Method to retrieve the complex coefficient \eqn{N}}
\usage{
N(obj)

\S4method{N}{PairCoefficients}(obj)
}
\arguments{
\item{obj}{Instance of \linkS4class{PairCoefficients}}
}
\description{
It is defined as \eqn{N = N_{11} + N_{10} + N_{01} + N_{00}} which equals 
\eqn{n \choose{2}} with \eqn{n} the number of objects
}
\author{
Fabian Ball \email{fabian.ball@kit.edu}
}
