% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\name{N00}
\alias{N00}
\alias{N00,PairCoefficients-method}
\title{Method to retrieve the coefficient \eqn{N_{00}}}
\usage{
N00(obj)

\S4method{N00}{PairCoefficients}(obj)
}
\arguments{
\item{obj}{Instance of \linkS4class{PairCoefficients}}
}
\description{
Method to retrieve the coefficient \eqn{N_{00}}
}
\author{
Fabian Ball \email{fabian.ball@kit.edu}
}
