% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\docType{methods}
\name{N01}
\alias{N01}
\alias{N01,PairCoefficients-method}
\title{Method to retrieve the coefficient \eqn{N_{01}}}
\usage{
N01(obj)

\S4method{N01}{PairCoefficients}(obj)
}
\arguments{
\item{obj}{Instance of \linkS4class{PairCoefficients}}
}
\description{
Method to retrieve the coefficient \eqn{N_{01}}
}
\author{
Fabian Ball \email{fabian.ball@kit.edu}
}
