% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\name{N01p}
\alias{N01p}
\alias{N01p,PairCoefficients-method}
\title{Method to retrieve the complex coefficient \eqn{N'_{01}}}
\usage{
N01p(obj)

\S4method{N01p}{PairCoefficients}(obj)
}
\arguments{
\item{obj}{Instance of \linkS4class{PairCoefficients}}
}
\description{
It is defined as \eqn{N'_{01} = N_{00} + N_{01}}
}
\author{
Fabian Ball \email{fabian.ball@kit.edu}
}
