% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\name{N10}
\alias{N10}
\alias{N10,PairCoefficients-method}
\title{Method to retrieve the coefficient \eqn{N_{10}}}
\usage{
N10(obj)

\S4method{N10}{PairCoefficients}(obj)
}
\arguments{
\item{obj}{Instance of \linkS4class{PairCoefficients}}
}
\description{
Method to retrieve the coefficient \eqn{N_{10}}
}
\author{
Fabian Ball \email{fabian.ball@kit.edu}
}
