% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\name{N10p}
\alias{N10p}
\alias{N10p,PairCoefficients-method}
\title{Method to retrieve the complex coefficient \eqn{N'_{10}}}
\usage{
N10p(obj)

\S4method{N10p}{PairCoefficients}(obj)
}
\arguments{
\item{obj}{Instance of \linkS4class{PairCoefficients}}
}
\description{
It is defined as \eqn{N'_{10} = N_{00} + N_{10}}
}
\author{
Fabian Ball \email{fabian.ball@kit.edu}
}
