% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\name{N11}
\alias{N11}
\alias{N11,PairCoefficients-method}
\title{Method to retrieve the coefficient \eqn{N_{11}}}
\usage{
N11(obj)

\S4method{N11}{PairCoefficients}(obj)
}
\arguments{
\item{obj}{Instance of \linkS4class{PairCoefficients}}
}
\description{
Method to retrieve the coefficient \eqn{N_{11}}
}
\author{
Fabian Ball \email{fabian.ball@kit.edu}
}
