% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\docType{methods}
\name{N12}
\alias{N12}
\alias{N12,PairCoefficients-method}
\title{Method to retrieve the complex coefficient \eqn{N_{12}}}
\usage{
N12(obj)

\S4method{N12}{PairCoefficients}(obj)
}
\arguments{
\item{obj}{Instance of \linkS4class{PairCoefficients}}
}
\description{
It is defined as \eqn{N_{12} = N_{11} + N_{01}}
}
\author{
Fabian Ball \email{fabian.ball@kit.edu}
}
