% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\name{N21}
\alias{N21}
\alias{N21,PairCoefficients-method}
\title{Method to retrieve the complex coefficient \eqn{N_{21}}}
\usage{
N21(obj)

\S4method{N21}{PairCoefficients}(obj)
}
\arguments{
\item{obj}{Instance of \linkS4class{PairCoefficients}}
}
\description{
It is defined as \eqn{N_{21} = N_{11} + N_{10}}
}
\author{
Fabian Ball \email{fabian.ball@kit.edu}
}
