% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R
\docType{class}
\name{Partition-class}
\alias{Partition-class}
\alias{Partition}
\title{Simple S4 class to represent a partition of objects as vector of class labels.}
\description{
This class is a wrapper around a vector but allows only the atomic vectors
logical, numeric, integer, complex, character, raw.
The reason for this is that only those types seem to make sense as class labels.
Furthermore, class labels are immutable.
}
\examples{
p <- new("Partition", c(0, 0, 1, 1, 1))
q <- new("Partition", c("a", "a", "b", "b", "b"))

\dontrun{
# This won't work:
new("Partition", c(list("a"), "a", "b", "b", "b"))
p[2] <- 2
}

}
\author{
Fabian Ball \email{fabian.ball@kit.edu}
}
