% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aab-helper.R
\name{projectionNumber}
\alias{projectionNumber}
\title{Compute the projection number of two partitions}
\usage{
projectionNumber(p, q)
}
\arguments{
\item{p}{Partition \eqn{P}}

\item{q}{Partition \eqn{Q}}
}
\description{
Given two partitions (p, q) represented as vectors of cluster ids,
compute the projection number which is the sum of maximum
cluster overlaps for all clusters of \eqn{P} to any cluster of \eqn{Q}.
}
\examples{
projectionNumber(c(0, 0, 0, 1, 1), c(0, 0, 1, 1, 1)) == 4

}
\seealso{
\code{\link{dongensMetric}}
}
\author{
Fabian Ball \email{fabian.ball@kit.edu}
}
